requirements-dev.txt
-r requirements.txt
black
mypy
pylint==2.13.9
PyJWT
cryptography
python-dotenv
python-socketio[client]
requests
types-requests